refactor: rename NumbersLink component
