+++ components/NumberLinks/index.ts
@@ -1,0 +1,1 @@
+export { default as NumberLinks } from "./NumberLinks";
